Dynamic Onboarding Questions
I added a new entity to keep the onboarding questions in the database. The entity comes with its data access object and repository to help with the CRUD functions.

I added a dynamic cubit and state to control the UI elements of the dynamic page.

I have modified the object box model to save the onboarding entity.

I added a new JSON object in the assets to act as the AWS connection

I have modified the avatar background to improve the look