Update from Backend 1
The Hub of the app is covered by a cubit consumer that checks for updates when the app is back in focus. The app then clears all the diaries that have not been started and pulls the new data. the app compares diaries to avoid duplication.

Made the payload nullable in the notification manager to avoid null exceptions