Milestone Six Final (#494)
* Clean up

* Dynamic Retrieval Updates

TODO: Stop the overriding of diaries

* Update from Backend 1

The Hub of the app is covered by a cubit consumer that checks for updates when the app is back in focus. The app then clears all the diaries that have not been started and pulls the new data. the app compares diaries to avoid duplication.

Made the payload nullable in the notification manager to avoid null exceptions

* Updated Modal

Updated the design for the update modal

* Update Study

1. Added The Onboarding details to the settings page
2. Added the Onboarding altering page to allow users to alter the details to change the study
3. Added new onboarding question type (date) to allow participants to select days as answers
4. Altered the onboarding to accommodate the new date picker

* Update From Main

* M6 Retool Changes

Modification to accommodate the retool changes

* Bug Fix

1. WebView Survey Completion - Detects on Android now
2. Timer - Clicking on the time's up stops the alarm, and adding/subtracting to the already started timer appends the remaining time, and clicking continue will shut off the alarm
3. AudioText - can delete the text response
4. Efficiently skipping the instructions question on the summary page
5. recognising the gesture for updating the study

* Update pubspec.yaml

Updating the version

* Removed Print Statements