workflow and template updates (#645)
* delete one workflow and kept one

Deleted the workflow which runs the MacOS since the Linux one does the same thing. This will free up our monthly allocated resources. And deleted the run on push to branches and kept the Pull Request run since we created Pull Requests when we are pushing changes to the repo. these changes will speed up the workflow.

* Rename of the test filename for clarity

* added code quality checks and templates

* Update UnitAndWidgetTesting.yml

* updated dart format checker