Dev (#628)
* 8/12 tests pass

* diaries page test

* settings page test

* tidy-up: diaries page test relocation

* rive_common_plugin error, added a fixme comment

created a helper function to help with the mock up for the rive animation testing. it is still not working. added a fixme comment and commented out the failing test cases

* Update diariespage_test.dart

* tidy-up: settings_test.dart

* tidy-up: homepage_test.dart

* added more homepage_test.dart tests and removed the widget_test.dart file

* added a github action workflow file for the running tests

* tidy up

* tidyup

* update to the workflow

* update to the workflow

* Update flutter_test.yml

* Update flutter_test.yml

* workflow rename

* dummy keys

* Update testing.yml

* renamed Preview to MediaPreview

* Added the action workflow for iOS and renamed the other workflow to androidTesting

re

* Integration testing implementation (#613)

* Instructions to run the tests in the README.md

* Integration testing added

* added tests details

* Update README.md

* Incentive pop added.

* Update README.md

* updated the onboarding flow and added a history flow

* settings page added in

* settings flow test done

* full E2E test

* set calendar date to dynamic in the integration test

* Tidy up

1: removed unused test code
2: removed patrol dependency, it is not used in testing
3: used the text continue to find the CustomFlatButton to reduce chances of missed tap hits by the tester

* added a default Pull Request Template

* Update pull_request_template.md

* Update README.md

* updated tests flows in integration test to reflect the onboarding update modal pop up

* Update README.md

* Unit tests all layers (#640)

* Unit tests for all layers

* Update pubspec.lock

* refactored widget and Integration test

added reusable methods in the widget tests and for the integration tests moved the flow tests into their own files for better readability and debugging as well as making it less tightly coupled.

* Refactored unit tests  by removing hardcoded test data for centralized dummy data functions for consistency across tests.

* Refactored Workflow (#643)

* delete one workflow and kept one

Deleted the workflow which runs the MacOS since the Linux one does the same thing. This will free up our monthly allocated resources. And deleted the run on push to branches and kept the Pull Request run since we created Pull Requests when we are pushing changes to the repo. these changes will speed up the workflow.

* Rename of the test filename for clarity

* workflow and template updates (#645)

* delete one workflow and kept one

Deleted the workflow which runs the MacOS since the Linux one does the same thing. This will free up our monthly allocated resources. And deleted the run on push to branches and kept the Pull Request run since we created Pull Requests when we are pushing changes to the repo. these changes will speed up the workflow.

* Rename of the test filename for clarity

* added code quality checks and templates

* Update UnitAndWidgetTesting.yml

* updated dart format checker

* added missing parameters for optional on formatter and respondedAt for promptEntry

---------

Co-authored-by: Kelvin Sande <98394184+sande11@users.noreply.github.com>